refactor: add support to locale resources

+++ backend/src/modules/emergency/usecases/zona/DeleteZonaUseCase.ts
@@ -1,5 +1,7 @@
+import { ErrorMessages } from '@/core/helpers/ErrorMessages'
 import { ZonaRepository } from '@/emergency/repositories/ZonaRepository '
-import { Injectable } from '@nestjs/common'
+import { Localidade } from '@/locality/structures/Localidade'
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common'
 
 import { DeleteResult } from 'typeorm'
 
@@ -9,7 +11,16 @@ export class DeleteZonaUseCase {
     private readonly zonaRepository: ZonaRepository,
   ) { }
 
-  async execute(id: number): Promise<DeleteResult> {
+  async execute(localidade: Localidade, id: number): Promise<DeleteResult> {
+    const model = await this.zonaRepository.findById(id)
+    if (!model) {
+      throw new NotFoundException(ErrorMessages.emergency.zona.notFound)
+    }
+
+    if (localidade.id !== model.localidadeId) {
+      throw new ForbiddenException(ErrorMessages.emergency.localidade.notAllowed)
+    }
+
     return this.zonaRepository.delete(id)
   }
 }
